requirements.txt
aiogram==3.13.1
groq==0.11.0
requests==2.32.3
python-dotenv==1.0.1
